# backend/ACG/anicomgo-backend/docs/apifox-openapi.yaml
openapi: 3.0.3
info:
  title: AniComGo 后端接口文档
  version: 1.4.0
  description: |
    AniComGo 后端 OpenAPI 文档，可直接导入 Apifox。
    所有接口统一使用 Result 响应结构：{ code, message, data }。
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      title: 用户登录请求
      description: 用户登录时提交的用户名和密码
      required: [username, password]
      properties:
        username:
          type: string
          description: 登录用户名
          maxLength: 50
        password:
          type: string
          description: 登录密码
          maxLength: 100
    RegisterRequest:
      type: object
      title: 用户注册请求
      description: 用户注册时提交的基础资料
      required: [username, password]
      properties:
        username:
          type: string
          description: 注册用户名，长度 3 到 50 个字符
          minLength: 3
          maxLength: 50
        password:
          type: string
          description: 注册密码，长度 6 到 100 个字符
          minLength: 6
          maxLength: 100
        avatarUrl:
          type: string
          description: 用户头像地址，可选；建议先调用通用上传接口 `/upload`，再将返回的完整图片 URL 填入该字段
          maxLength: 255
        bio:
          type: string
          description: 用户个人简介，可选
          maxLength: 300
...
    CollectionCreateRequest:
      type: object
      title: 新增收藏请求
      description: 用户收藏番剧时提交的番剧、进度和评分信息
      required: [animeId]
      properties:
        animeId:
          type: integer
          format: int64
          description: 要收藏的番剧 ID
          minimum: 1
        progress:
          type: integer
          description: 当前观看进度，可选
          minimum: 0
        rating:
          type: integer
          description: 当前评分，可选，范围 1 到 5
          minimum: 1
          maximum: 5
    UpdateCollectionProgressRequest:
      type: object
      title: 修改追番进度请求
      description: 用户修改自己某部已收藏番剧的当前追番进度
      required: [progress]
      properties:
        progress:
          type: integer
          description: 最新追番进度，不能小于 0，且不能超过该番剧总集数（如果已设置总集数）
          minimum: 0
...
  responses:
    SuccessAuth:
      description: 认证成功
      content:
        application/json:
          example:
            code: 200
            message: 登录成功
            data:
              token: eyJhbGciOiJIUzI1NiJ9.demo.token
              user:
                id: 1
                username: 御坂10032
                avatarUrl: https://cdn.example.com/avatar/misaka.jpg
                bio: 超电磁炮爱好者
                role: 0
                status: 1
    SuccessProfile:
      description: 用户资料查询成功
      content:
        application/json:
          example:
            code: 200
            message: 成功
            data:
              id: 1
              username: 御坂10032
              avatarUrl: https://cdn.example.com/avatar/misaka.jpg
              bio: 超电磁炮爱好者
              role: 0
              status: 1
...
    SuccessCollectionPage:
      description: 收藏分页查询成功
      content:
        application/json:
          example:
            code: 200
            message: 成功
            data:
              records:
                - id: 5
                  userId: 1
                  animeId: 1
                  progress: 12
                  rating: 5
              total: 1
              size: 10
              current: 1
...
    Unauthorized:
      description: 未登录、令牌无效或登录状态失效
      content:
        application/json:
          example:
            code: 401
            message: 请先登录
            data: null
    Forbidden:
      description: 无权限访问
      content:
        application/json:
          example:
            code: 403
            message: 无权限访问
            data: null
...
paths:
  /auth/register:
    post:
      tags: [用户端-认证]
      summary: 用户注册
      description: 创建普通用户账号，可选提交头像地址和个人简介。注册成功后直接返回令牌与用户信息。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: misaka_10032
              password: 12345678
              avatarUrl: http://127.0.0.1:9000/avatar/2026/default-user.jpg
              bio: 超电磁炮爱好者
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              example:
                code: 200
                message: 注册成功
                data:
                  token: eyJhbGciOiJIUzI1NiJ9.demo.token
                  user:
                    id: 1
                    username: misaka_10032
                    avatarUrl: http://127.0.0.1:9000/avatar/2026/default-user.jpg
                    bio: 超电磁炮爱好者
                    role: 0
                    status: 1
        '400':
          description: 参数校验失败或用户名已存在
          content:
            application/json:
              example:
                code: 400
                message: 用户名已存在
                data: null
  /auth/login:
    post:
      tags: [用户端-认证]
      summary: 用户登录
      description: 使用用户名和密码登录系统。登录成功后返回 JWT 令牌，前端后续访问受保护接口时需携带该令牌。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: misaka_10032
              password: 12345678
      responses:
        '200':
          $ref: '#/components/responses/SuccessAuth'
        '401':
          description: 用户名或密码错误
          content:
            application/json:
              example:
                code: 401
                message: 用户名或密码错误
                data: null
        '403':
          description: 账号已被禁用
          content:
            application/json:
              example:
                code: 403
                message: 账号已被禁用
                data: null
  /auth/me:
    get:
      tags: [用户端-认证]
      summary: 获取当前登录用户信息
      description: 根据当前请求携带的 JWT，返回当前登录用户的公开资料，用于前端初始化登录态。
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          $ref: '#/components/responses/SuccessProfile'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: 当前用户不存在
          content:
            application/json:
              example:
                code: 404
                message: 用户不存在
                data: null
...
  /collections:
    get:
      tags: [用户端-收藏]
      summary: 分页查询收藏列表
      description: 面向用户端的收藏列表查询接口，可按用户 ID 查看其收藏记录。
      parameters:
        - in: query
          name: page
          description: 页码，从 1 开始
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - in: query
          name: size
          description: 每页条数
          required: false
          schema:
            type: integer
            minimum: 1
            default: 5
          example: 5
        - in: query
          name: userId
          description: 按用户 ID 查询其收藏记录
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 1
      responses:
        '200':
          $ref: '#/components/responses/SuccessCollectionPage'
    post:
      tags: [用户端-收藏]
      summary: 新增收藏
      description: 当前登录用户收藏某部番剧，并可同步记录观看进度和评分。
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreateRequest'
            example:
              animeId: 1
              progress: 12
              rating: 5
      responses:
        '200':
          description: 新增收藏成功
          content:
            application/json:
              example:
                code: 200
                message: 新增收藏成功
                data: null
        '400':
          description: 重复收藏或评分、进度参数不合法
          content:
            application/json:
              example:
                code: 400
                message: 收藏记录已存在
                data: null
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: 番剧不存在
          content:
            application/json:
              example:
                code: 404
                message: 番剧不存在
                data: null
  /collections/anime/{animeId}:
    get:
      tags: [用户端-收藏]
      summary: 查询当前用户对指定番剧的收藏详情
      description: 当前登录用户根据番剧 ID 查询自己的收藏记录，可用于番剧收藏页回显当前评分、追番进度等信息。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: animeId
          description: 番剧 ID
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
          example: 1
      responses:
        '200':
          description: 查询收藏详情成功
          content:
            application/json:
              example:
                code: 200
                message: 成功
                data:
                  id: 5
                  userId: 1
                  animeId: 1
                  progress: 12
                  rating: 5
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: 番剧不存在或当前用户未收藏该番剧
          content:
            application/json:
              examples:
                番剧不存在:
                  value:
                    code: 404
                    message: 番剧不存在
                    data: null
                收藏不存在:
                  value:
                    code: 404
                    message: 收藏不存在
                    data: null
  /collections/anime/{animeId}/progress:
    patch:
      tags: [用户端-收藏]
      summary: 修改当前用户对指定番剧的追番进度
      description: 当前登录用户修改自己已收藏番剧的追番进度。适用于收藏页或详情页中的“看到第几集”更新操作。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: animeId
          description: 番剧 ID
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCollectionProgressRequest'
            example:
              progress: 13
      responses:
        '200':
          description: 修改追番进度成功
          content:
            application/json:
              example:
                code: 200
                message: 修改追番进度成功
                data:
                  id: 5
                  userId: 1
                  animeId: 1
                  progress: 13
                  rating: 5
        '400':
          description: 进度不合法或超过总集数
          content:
            application/json:
              examples:
                进度超出总集数:
                  value:
                    code: 400
                    message: 观看进度不能超过总集数
                    data: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: 番剧不存在或当前用户未收藏该番剧
          content:
            application/json:
              examples:
                番剧不存在:
                  value:
                    code: 404
                    message: 番剧不存在
                    data: null
                收藏不存在:
                  value:
                    code: 404
                    message: 收藏不存在
                    data: null
  /collections/{id}:
    delete:
      tags: [用户端-收藏]
      summary: 删除收藏
      description: 当前登录用户删除自己的收藏记录；管理员也可执行该操作。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          description: 收藏记录 ID
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
          example: 5
      responses:
        '200':
          description: 删除收藏成功
          content:
            application/json:
              example:
                code: 200
                message: 删除收藏成功
                data: null
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: 仅收藏所属用户本人或管理员可操作
          content:
            application/json:
              example:
                code: 403
                message: 你只能操作自己的收藏
                data: null
        '404':
          description: 收藏不存在
          content:
            application/json:
              example:
                code: 404
                message: 收藏不存在
                data: null
...
  /admin/collections/{id}:
    delete:
      tags: [管理端-收藏管理]
      summary: 删除任意收藏
      description: 仅管理员可用。根据收藏 ID 删除任意收藏记录。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          description: 收藏记录 ID
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
          example: 5
      responses:
        '200':
          description: 删除收藏成功
          content:
            application/json:
              example:
                code: 200
                message: 管理员删除收藏成功
                data: null
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: 收藏不存在
          content:
            application/json:
              example:
                code: 404
                message: 收藏不存在
                data: null
...